big fix

+++ BeeStore_backend_library/DTO/OrderDTOs/OrderCreateDTO.cs
@@ -36,7 +36,7 @@ namespace BeeStore_Repository.DTO.OrderDTOs
 
         public int? LotId { get; set; }
         [JsonIgnore]
-        public int? ProductPrice { get; set; }
+        public decimal? ProductPrice { get; set; }
 
         public int? ProductAmount { get; set; }
     }
